bindings CHANGE tests refactored

+++ bindings/python/tests/MoveTest.py
@@ -22,14 +22,35 @@ import sysrepo as sr
 
 class MoveTest(unittest.TestCase):
 
-    @classmethod
-    def tearDown(self):
+
+    def remove_modules(self):
         TestModule.remove_test_module()
         TestModule.remove_referenced_data_module()
 
+    @classmethod
+    def setUpClass(self):
+        self.remove_modules(self)
+
+    @classmethod
+    def tearDownClass(self):
+        self.remove_modules(self)
+
+    @classmethod
+    def tearDown(self):
+        self.remove_modules(self)
+
+    @classmethod
     def setUp(self):
-        TestModule.create_referenced_data_module()
-        TestModule.create_test_module()
+        if not TestModule.create_referenced_data_module():
+            self.remove_modules(self)
+            self.skipTest(self,"Test environment is not clean!")
+            print("Environment is not clean!")
+            return
+        if not TestModule.create_test_module():
+            self.remove_modules(self)
+            self.skipTest(self,"Test environment is not clean!")
+            print("Environment is not clean!")
+            return
         conn = sr.Connection(sr.SR_CONN_DEFAULT)
         session = sr.Session(conn, sr.SR_DS_STARTUP)
         session.delete_item("/test-module:user[name='A']")
